link with libfvwm3.a to use fvwm gradients
* hsetroot.c: add a new option: -fvwm-gradient which uses fvwm3 code
to produce a gradient pixmap according to a gradient spec, and render
that on root window. See "Color Gradients" in man
fvwm3commands(1). e.g.
```
  ./hsetroot -fvwm-gradient "CGradient 128 #00b0b0 #006060"
```

* Makefile: make configurable. call `make LINK_FVWM3=0' to skip
linking to libfvwm3.a otherwise supply a config-link-fvwm3.mk which
sets up CFLAGS, LDFLAGS and FVW3LIB_LIB which points to libfvwm3.a.
default to LINK_FVWM3=1, uses "(make) Target-specific" variables.

* config-link-fvwm3.mk: configure the path to libfvwm.a